requirements.txt
anthropic==0.77.0
fastapi==0.128.0
pydantic==2.12.5
python-dotenv==1.2.1
Requests==2.32.5
supabase==2.27.2
uvicorn==0.40.0
